refactor : mainpage 코드 정리, 리팩토링
