+++ src/components/main/mainTitle/components/TitleBackground.jsx
@@ -6,12 +6,13 @@ const TitleBackground = () => {
 };
 
 const Container = styled.div`
-  background: ${({ theme }) => theme.subColor1};
-  width: 100vw;
-  height: 55.1rem;
   position: absolute;
   top: 0;
   left: 0;
+  width: 100vw;
+  height: 55.1rem;
+
+  background: ${({ theme }) => theme.subColor1};
   border-radius: 0px 0px 0px 142px;
   z-index: -2;
 `;
