+++ src/components/main/mainTitle/screen/MainTitleScreen.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+
 import PageTitle from "../components/PageTitle";
 import TitleBackground from "../components/TitleBackground";
